Fixed all comments of my tutor
