+++ js/main.js
@@ -4,3 +4,4 @@ import './user-miniature.js';
 import './render-big-picture.js';
 import './upload-picture.js';
 import './validate-form.js';
+import './show-success-error-popup.js';
