+++ js/upload-picture.js
@@ -1,7 +1,6 @@
 /* noUiSlider */
-import {isEscEvent} from './util.js';
 import {hashTagInput} from './validate-form.js';
-import api from './api.js';
+import {isEscEvent} from './util.js';
 
 let scale = 100;
 const percent = '%';
@@ -28,8 +27,6 @@ const effectLevel = editingForm.querySelector('.img-upload__effect-level');
 const effectLevelSlider = editingForm.querySelector('.effect-level__slider');
 const uploadPreviewImg = editingForm.querySelector('.img-upload__preview').querySelector('img');
 const effectLevelValue = editingForm.querySelector('.effect-level__value');
-const imgUploadForm = document.querySelector('.img-upload__form');
-const successTemplate = document.querySelector('#success').content.querySelector('.success');
 
 const onCloseEditingFormClick = () => {
   editingForm.classList.add('hidden');
@@ -46,7 +43,6 @@ const onCloseEditingFormEscKeydown = (evt) => {
     evt.preventDefault();
     editingForm.classList.add('hidden');
     document.body.classList.remove('modal-open');
-    hashTagInput.style.border = 'none';
   }
 };
 uploadInput.addEventListener('change', (evt) => {
@@ -139,78 +135,9 @@ effectLevelSlider.noUiSlider.on('change', () => {
   uploadPreviewImg.style.filter = effects[lastClass.replace('effects__preview--', '')]();
 });
 
-const onSuccesButtonClick = (evt) => {
-  if (evt.currentTarget) {
-    const successSection = document.querySelector('.success');
-    successSection.remove();
-  }
-}
-
-const onSuccesButtonEscKeydown = (evt) => {
-  if (isEscEvent(evt)) {
-    const successSection = document.querySelector('.success');
-    successSection.remove();
-    document.body.removeEventListener('keydown', onSuccesButtonEscKeydown);
-  }
-}
-
-const onErrorButtonClick = (evt) => {
-  if (evt.currentTarget) {
-    const errorSection = document.querySelector('.error');
-    errorSection.remove();
-  }
-}
-
-const onErrorButtonEscKeydown = (evt) => {
-  if (isEscEvent(evt)) {
-    const errorSection = document.querySelector('.error');
-    errorSection.remove();
-    document.body.removeEventListener('keydown', onErrorButtonEscKeydown);
-  }
-}
-
-imgUploadForm.addEventListener('submit', (evt) => {
-  evt.preventDefault();
-  const formData = new FormData(imgUploadForm);
-  api.postData(formData).then((response) => {
-    if (response) {
-      const successTemplateBlock = successTemplate.cloneNode(true);
-      document.body.appendChild(successTemplateBlock);
-      const successDiv = successTemplateBlock.querySelector('.success__inner');
-      successDiv.setAttribute('tabindex', '1');
-      successDiv.style.outline = 'none';
-      successDiv.focus();
-      editingForm.classList.add('hidden');
-      document.body.classList.remove('modal-open');
-      uploadInput.value = '';
-      uploadPreviewImg.style.transform = 'scale(1)';
-      uploadPreviewImg.style.filter = 'none';
-      hashTagInput.style.border = 'none';
-      const successButton = successTemplateBlock.querySelector('.success__button');
-      successButton.addEventListener('click', onSuccesButtonClick);
-      document.body.addEventListener('keydown', onSuccesButtonEscKeydown);
-      successDiv.addEventListener('blur', onSuccesButtonClick);
-      // document.addEventListener('click', onSuccesButtonClick);
-      // закрыть, сбросить форму
-      // навесить события
-    }
-  }).catch(() => {
-    const errorPostTemplate = document.querySelector('#error').content.querySelector('.error');
-    const errorPostTemplateBlock = errorPostTemplate.cloneNode(true);
-    document.body.appendChild(errorPostTemplateBlock);
-    const errorDiv = errorPostTemplateBlock.querySelector('.error__inner');
-    errorDiv.setAttribute('tabindex', '1');
-    errorDiv.style.outline = 'none';
-    errorDiv.focus();
-    editingForm.classList.add('hidden');
-    document.body.classList.remove('modal-open');
-    const errorButton = errorPostTemplateBlock.querySelector('.error__button');
-    errorButton.addEventListener('click', onErrorButtonClick);
-    document.body.addEventListener('keydown', onErrorButtonEscKeydown);
-    errorDiv.addEventListener('blur', onErrorButtonClick);
-
-    // навесить события
-  });
-});
-
-export {editingForm};
+export {
+  editingForm,
+  uploadInput,
+  uploadPreviewImg,
+  hashTagInput
+};
